feat: The basic bounded latency scheme
Cuju’s output buffering mechanism introduces the extra latency in the
epoch-based FT scheme. The length of the latency penalty is variable and
potentially unbounded because of the variation in the length of the run
stages and transfer stages. This variable and unbounded latency penalty
are unacceptable for certain applications that promise bounded latency
to their users. This commit provides a bounded latency version of the FT
system to support any highly latency-sensitive network applications in
VM.

The application latency as the sum of each epoch’s first three stages:
run, snapshot and transfer stages. The system could control the length
of a run stage by terminating it at an appropriate time. When the run
stage was terminated, the snapshot and transfer stage start. The length
of the run stage will impact the length of the transfer stage. However,
the system could not do much to dictate the length of a transfer stage
except by limiting the number of dirty bytes to be transferred.

The length of the transfer stage is typically variable for two reasons:
(a) variation in the total dirty byte count due to variation in dirty
page count and compression gain, and (b) variation in the transfer rate.
It can be seen that the main challenge to bound the application latency
is how to accurately predict the state transfer time required for a set
of dirty pages accumulated during a run stage; furthermore, terminate
the run stage at the right moment so that the resulting sum of the run,
snapshot and transfer stage falls below the target latency. Prediction
of the transfer stage’s length for a given set of dirty pages is
non-trivial because dirty pages are compressed before being transferred,
and requires extra measurements from previous and current epochs. These
measurements may incur additional overheads and thus should be reduced
to the minimum.

This commit provides the bounded latency feature including the design as
follow:

1. Scout dirty byte counts
The thread scout dirty byte counts and pin in a CPU core. So that it can
calculate the dirty byte counts in the background and doesn't incur the
extra overhead of guest VM.

2. The frequency of sampling in the dirty List
To avoid performance impairment, try to reduce the times of the dirty
pages comparison. Skip m pages and then do the comparison of n dirty
pages. Finally get the dirty byte counts and scale to the whole
distribution of dirty page counts to speculative the real dirty byte
counts.

3. Speculative dirty byte counts
Each internal time between calculating dirty byte counts from the dirty
list, we can get the dirty byte counts rate, which is the increment of
dirty byte per microsecond. We get the extra dirty bytes from taking the
last dirty byte counts rate to multiply the time of calculating dirty
byte counts in the current sampling. The last dirty byte counts rate try
to compensate for the possible new dirty byte counts produce during the
current sampling.

4. Predict variable transfer rate
Transfer time is composed of the time taken to transfer snapshot states
to the backup machine and to receive ACKs from the backup machine.
During transfer stages, the transfer thread compresses dirty page one by
one and sends out the compress data to its backup machine. After sending
out all snapshot states, the transfer thread must wait for ACKs from the
slave VM, thus the master VM, to be sure that all states are in backup
machines memory. Therefore, the transfer time also includes the time to
load received data in slave VMs memory on the backup machine. Also,
communications between the master and slave VMs are also counted into
total transfer time. Another main factor that affects the transfer rate
is the TCP status on both master and slave VM. The window sizes of TCP
stacks on both sides need some time to grow. This is reflected in recent
historical transfer rates. Therefore we record the real transfer rate
for each epoch and average the last 5 epochs' transfer rate as the next
prediction of epoch's transfer rate.

5. When to take the snapshot
In the run stage, the thread in the background collects the related
information, including the current epoch's runtime (E), anticipative
transfer rate (R) and dirty byte counts (B). When E+B/R >
target_latency, it must enter the snapshot stage.

6. Latency acceptance
If we set the target latency as 10 ms, the range of latency acceptance
was defined as 9ms to 11ms. The accuracy of bounded latency represents
how many epochs' latency in the range of latency acceptance (9~11ms)
within total epochs' counts. The bounded latency of this commit can
achieve 96% accuracy while compiling linux kernel source in guest VM.